requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
astunparse==1.6.3
beautifulsoup4==4.12.3
binaryornot==0.4.4
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cookiecutter==2.6.0
fastapi==0.115.2
flatbuffers==24.3.25
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
frozendict==2.4.6
gast==0.6.0
google-ai-generativelanguage==0.4.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.62.3
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
keras==2.15.0
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
ml-dtypes==0.2.0
multitasking==0.0.11
namex==0.0.8
neo4j-driver==5.25.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.0
packaging==23.2
pandas==2.2.3
peewee==3.17.7
pip-check-reqs==2.5.3
platformdirs==4.3.6
proto-plus==1.24.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.9.11
repath==0.9.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.1
starlette==0.40.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
termcolor==2.5.0
text-unidecode==1.3
tqdm==4.66.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchdog==4.0.2
watchfiles==0.24.0
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.4
wrapt==1.14.1
yfinance==0.2.44
